benchmarks: automated report generation
