+++ benchmarks/report.js
@@ -1,20 +1,17 @@
 'use strict'
 
-const concat = require('concat-stream')
-const handlebars = require('handlebars')
-const fs = require('fs')
 const join = require('path').join
-const template = handlebars.compile(
-  fs.readFileSync(join(__dirname, 'report-template', 'report.html'), 'utf8'))
+const exec = require('child_process').exec
+const open = require('opn')
 
-process.stdin.pipe(concat(gotResult))
-
-function gotResult(str) {
-  const result = JSON.parse(str)
-  outputReport(result)
-}
-
-function outputReport(result) {
-  const html = template({ results: result })
-  process.stdout.write(html)
-}
+const outDir = join(__dirname, 'reports', 'out')
+const out = join(outDir, Date.now() + '-report.html')
+const command = 'node benchmarks | node benchmarks/generate-report > ' + out
+const child = exec(command)
+child.stderr.pipe(process.stderr)
+child.once('exit', (code, signal) => {
+  if (code === 0) {
+    process.stderr.write('Report generated to ' + out + '\n')
+    open(out, { wait: false })
+  }
+})
