+++ package.json
@@ -42,7 +42,8 @@
     "release-minor": "gulp release --type minor",
     "release-major": "gulp release --type major",
     "coverage-publish": "aegir-coverage publish",
-    "benchmarks": "node benchmarks"
+    "benchmarks": "node benchmarks",
+    "benchmarks:report": "node benchmarks/report"
   },
   "pre-commit": [
     "lint",
@@ -85,6 +86,7 @@
     "mocha": "^3.2.0",
     "ncp": "^2.0.0",
     "nexpect": "^0.5.0",
+    "opn": "^4.0.2",
     "platform": "^1.3.3",
     "pre-commit": "^1.2.2",
     "qs": "^6.3.0",
